+++ src/hooks/useAnimateFade.tsx
@@ -1,0 +1,30 @@
+import {useRef} from 'react';
+import {Animated} from 'react-native';
+
+const useAnimateFade = () => {
+  const opacity = useRef(new Animated.Value(0)).current;
+
+  const fadeIn = () => {
+    Animated.timing(opacity, {
+      toValue: 1,
+      duration: 300,
+      useNativeDriver: true,
+    }).start();
+  };
+
+  const fadeOut = () => {
+    Animated.timing(opacity, {
+      toValue: 0,
+      duration: 300,
+      useNativeDriver: true,
+    }).start();
+  };
+
+  return {
+    opacity,
+    fadeIn,
+    fadeOut,
+  };
+};
+
+export default useAnimateFade;
